Remove pytest runtime dependency due to asdf tags for SkyCoord and TimeDelta
