+++ astropy/io/misc/asdf/tags/coordinates/skycoord.py
@@ -1,12 +1,9 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 # -*- coding: utf-8 -*-
 
-import numpy as np
+from asdf.yamlutil import custom_tree_to_tagged_tree
 
-from asdf.yamlutil import custom_tree_to_tagged_tree, tagged_tree_to_custom_tree
-
-from astropy.coordinates import SkyCoord
-from astropy.table.tests.test_operations import skycoord_equal
+from astropy.coordinates import SkyCoord, skycoord_equal
 
 from ...types import AstropyType
